Split src and include dirs.
