+++ torchvision/cinclude/torchvision/models/densenet.h
