+++ torchvision/cinclude/torchvision/models/inception.h
