+++ torchvision/cinclude/torchvision/models/squeezenet.h
