+++ torchvision/csrc/cuda/DeformConv_cuda.cu
@@ -72,7 +72,7 @@
 #include <c10/cuda/CUDAGuard.h>
 #include <THC/THCAtomics.cuh>
 
-#include "cuda_helpers.h"
+#include "cuda/cuda_helpers.h"
 
 #include <cmath>
 #include <iostream>
