+++ torchvision/csrc/cuda/ROIPool_cuda.cu
@@ -4,7 +4,7 @@
 #include <c10/cuda/CUDAGuard.h>
 #include <THC/THCAtomics.cuh>
 
-#include "cuda_helpers.h"
+#include "cuda/cuda_helpers.h"
 
 template <typename T>
 __global__ void RoIPoolForward(
