+++ torchvision/csrc/cuda/nms_cuda.cu
@@ -2,7 +2,7 @@
 #include <ATen/cuda/CUDAContext.h>
 #include <c10/cuda/CUDAGuard.h>
 
-#include "cuda_helpers.h"
+#include "cuda/cuda_helpers.h"
 
 #include <iostream>
 #include <vector>
